Global optimization and float64 type

+++ materials.py
@@ -1,6 +1,5 @@
-from numba import jit
 from numba.typed import List
-from numpy import array, load, unique, zeros, nonzero, interp, float32
+from numpy import array, load, unique, zeros, nonzero, interp
 
 mac_table = List([
     load('macs/Air, Dry (near sea level).npy'),      # 0 - Воздух
@@ -43,7 +42,7 @@ process_indices = {
 }
 
 def get_lac(material, energy, processes):
-    lac_out = zeros((len(processes), energy.size), dtype=float32)
+    lac_out = zeros((len(processes), energy.size))
     for m in unique(material):
         lac = lac_table[m]
         indices = nonzero(material == m)[0]
